+++ app.go
@@ -21,7 +21,7 @@ import (
 	"time"
 )
 
-var version = "0.8.3"
+var version = "0.8.4"
 
 const (
 	maxCachedQueries = 512